+++ packages/expo-modules-core/src/index.ts
@@ -5,13 +5,10 @@ import NativeModulesProxy from './NativeModulesProxy';
 import { ProxyNativeModule } from './NativeModulesProxy.types';
 import { requireNativeViewManager } from './NativeViewManagerAdapter';
 import Platform from './Platform';
-import SyntheticPlatformEmitter from './SyntheticPlatformEmitter';
 import { CodedError } from './errors/CodedError';
 import { UnavailabilityError } from './errors/UnavailabilityError';
 
 import './sweet/setUpErrorManager.fx';
-
-export { default as deprecate } from './deprecate';
 
 export { uuidv4, uuidv5 } from './uuid';
 
@@ -22,7 +19,6 @@ export {
   ProxyNativeModule,
   Platform,
   Subscription,
-  SyntheticPlatformEmitter,
   requireNativeViewManager,
   // Errors
   CodedError,
@@ -35,7 +31,7 @@ export * from './TypedArrays.types';
 /**
  * @deprecated renamed to `DeviceEventEmitter`
  */
-export const RCTDeviceEventEmitter = DeviceEventEmitter;
+export const SyntheticPlatformEmitter = DeviceEventEmitter;
 
 export * from './PermissionsInterface';
 export * from './PermissionsHook';